Create .paas directory and put stackato stuff in it
This is a bit of an experiment but I've created .paas/stackato and put
stackato.yml, app-store.json and README.md in it.

These are all stackato specific so it makes sense to do it in one place, and
in a place that the upstream would consider integrating.

I had to symlink stackato.yml to the top dir, but it all still works.